move GhostBatchNormHparams
